Finish refactors
