+++ lib/ain-evm/src/evm.rs
@@ -1,484 +1,263 @@
-use crate::backend::{EVMBackend, EVMBackendError, InsufficientBalance, Vicinity};
-use crate::block::INITIAL_BASE_FEE;
-use crate::executor::TxResponse;
-use crate::fee::calculate_prepay_gas;
+use crate::backend::{EVMBackend, Vicinity};
+use crate::block::BlockService;
+use crate::core::{EVMError, NativeTxHash, EVMCoreService, MAX_GAS_PER_BLOCK};
+use crate::executor::{AinExecutor, TxResponse};
+use crate::filters::FilterService;
+use crate::log::LogService;
 use crate::receipt::ReceiptService;
-use crate::storage::traits::{BlockStorage, PersistentStateError};
+use crate::storage::traits::BlockStorage;
 use crate::storage::Storage;
+use crate::traits::Executor;
 use crate::transaction::bridge::{BalanceUpdate, BridgeTx};
-use crate::trie::TrieDBStore;
-use crate::tx_queue::{QueueError, QueueTx, TransactionQueueMap};
-use crate::{
-    executor::AinExecutor,
-    traits::{Executor, ExecutorContext},
-    transaction::SignedTx,
-};
-use anyhow::anyhow;
-use ethereum::{AccessList, Account, Block, Log, PartialHeader, TransactionV2};
-use ethereum_types::{Bloom, BloomInput, H160, U256};
+use crate::trie::GENESIS_STATE_ROOT;
+use crate::txqueue::QueueTx;
 
-use hex::FromHex;
+use anyhow::anyhow;
+use ethereum::{Block, PartialHeader, ReceiptV3};
+use ethereum_types::{Bloom, H160, H64, U256};
 use log::debug;
+use primitive_types::H256;
 use std::error::Error;
 use std::path::PathBuf;
 use std::sync::Arc;
-use vsdb_core::vsdb_set_base_dir;
 
-pub type NativeTxHash = [u8; 32];
-
-pub const MAX_GAS_PER_BLOCK: U256 = U256([30_000_000, 0, 0, 0]);
-
-pub struct TxQueueService {
-    pub tx_queues: Arc<TransactionQueueMap>,
-    pub trie_store: Arc<TrieDBStore>,
-    storage: Arc<Storage>,
-}
-pub struct EthCallArgs<'a> {
-    pub caller: Option<H160>,
-    pub to: Option<H160>,
-    pub value: U256,
-    pub data: &'a [u8],
-    pub gas_limit: u64,
-    pub access_list: AccessList,
-    pub block_number: U256,
+pub struct EVMServices {
+    pub core: EVMCoreService,
+    pub block: BlockService,
+    pub receipt: ReceiptService,
+    pub logs: LogService,
+    pub filters: FilterService,
+    pub storage: Arc<Storage>,
 }
 
-fn init_vsdb() {
-    debug!(target: "vsdb", "Initializating VSDB");
-    let datadir = ain_cpp_imports::get_datadir().expect("Could not get imported datadir");
-    let path = PathBuf::from(datadir).join("evm");
-    if !path.exists() {
-        std::fs::create_dir(&path).expect("Error creating `evm` dir");
-    }
-    let vsdb_dir_path = path.join(".vsdb");
-    vsdb_set_base_dir(&vsdb_dir_path).expect("Could not update vsdb base dir");
-    debug!(target: "vsdb", "VSDB directory : {}", vsdb_dir_path.display());
-}
+pub type FinalizedBlockInfo = ([u8; 32], Vec<String>, u64);
 
-impl TxQueueService {
-    pub fn restore(storage: Arc<Storage>) -> Self {
-        init_vsdb();
-
-        Self {
-            tx_queues: Arc::new(TransactionQueueMap::new()),
-            trie_store: Arc::new(TrieDBStore::restore()),
-            storage,
+impl EVMServices {
+    /// Constructs a new Handlers instance. Depending on whether the defid -ethstartstate flag is set,
+    /// it either revives the storage from a previously saved state or initializes new storage using input from a JSON file.
+    /// This JSON-based initialization is exclusively reserved for regtest environments.
+    ///
+    /// # Warning
+    ///
+    /// Loading state from JSON will overwrite previous stored state
+    ///
+    /// # Errors
+    ///
+    /// This method will return an error if an attempt is made to load a genesis state from a JSON file outside of a regtest environment.
+    ///
+    /// # Return
+    ///
+    /// Returns an instance of the struct, either restored from storage or created from a JSON file.
+    pub fn new() -> Result<Self, anyhow::Error> {
+        if let Some(path) = ain_cpp_imports::get_state_input_json() {
+            if ain_cpp_imports::get_network() != "regtest" {
+                return Err(anyhow!(
+                    "Loading a genesis from JSON file is restricted to regtest network"
+                ));
+            }
+            let storage = Arc::new(Storage::new());
+            Ok(Self {
+                core: EVMCoreService::new_from_json(Arc::clone(&storage), PathBuf::from(path)),
+                block: BlockService::new(Arc::clone(&storage)),
+                receipt: ReceiptService::new(Arc::clone(&storage)),
+                logs: LogService::new(Arc::clone(&storage)),
+                filters: FilterService::new(),
+                storage,
+            })
+        } else {
+            let storage = Arc::new(Storage::restore());
+            Ok(Self {
+                core: EVMCoreService::restore(Arc::clone(&storage)),
+                block: BlockService::new(Arc::clone(&storage)),
+                receipt: ReceiptService::new(Arc::clone(&storage)),
+                logs: LogService::new(Arc::clone(&storage)),
+                filters: FilterService::new(),
+                storage,
+            })
         }
     }
 
-    pub fn new_from_json(storage: Arc<Storage>, path: PathBuf) -> Self {
-        debug!("Loading genesis state from {}", path.display());
-        init_vsdb();
+    pub fn finalize_block(
+        &self,
+        context: u64,
+        update_state: bool,
+        difficulty: u32,
+        beneficiary: H160,
+        timestamp: u64,
+    ) -> Result<FinalizedBlockInfo, Box<dyn Error>> {
+        let mut all_transactions = Vec::with_capacity(self.core.tx_queues.len(context));
+        let mut failed_transactions = Vec::with_capacity(self.core.tx_queues.len(context));
+        let mut receipts_v3: Vec<ReceiptV3> = Vec::with_capacity(self.core.tx_queues.len(context));
+        let mut gas_used = 0u64;
+        let mut logs_bloom: Bloom = Bloom::default();
 
-        let handler = Self {
-            tx_queues: Arc::new(TransactionQueueMap::new()),
-            trie_store: Arc::new(TrieDBStore::new()),
-            storage: Arc::clone(&storage),
-        };
-        let (state_root, genesis) =
-            TrieDBStore::genesis_state_root_from_json(&handler.trie_store, &handler.storage, path)
-                .expect("Error getting genesis state root from json");
-
-        let block: Block<TransactionV2> = Block::new(
-            PartialHeader {
-                state_root,
-                number: U256::zero(),
-                beneficiary: Default::default(),
-                receipts_root: ReceiptService::get_receipts_root(&Vec::new()),
-                logs_bloom: Default::default(),
-                gas_used: Default::default(),
-                gas_limit: genesis.gas_limit.unwrap_or(MAX_GAS_PER_BLOCK),
-                extra_data: genesis.extra_data.unwrap_or_default().into(),
-                parent_hash: genesis.parent_hash.unwrap_or_default(),
-                mix_hash: genesis.mix_hash.unwrap_or_default(),
-                nonce: genesis.nonce.unwrap_or_default(),
-                timestamp: genesis.timestamp.unwrap_or_default().as_u64(),
-                difficulty: genesis.difficulty.unwrap_or_default(),
-            },
-            Vec::new(),
-            Vec::new(),
-        );
-        storage.put_latest_block(Some(&block));
-        storage.put_block(&block);
-        // NOTE(canonbrother): set an initial base fee for genesis block
-        // https://github.com/ethereum/go-ethereum/blob/46ec972c9c56a4e0d97d812f2eaf9e3657c66276/params/protocol_params.go#LL125C2-L125C16
-        storage.set_base_fee(block.header.hash(), INITIAL_BASE_FEE);
-
-        handler
-    }
-
-    pub fn flush(&self) -> Result<(), PersistentStateError> {
-        self.trie_store.flush()
-    }
-
-    pub fn call(&self, arguments: EthCallArgs) -> Result<TxResponse, Box<dyn Error>> {
-        let EthCallArgs {
-            caller,
-            to,
-            value,
-            data,
-            gas_limit,
-            access_list,
-            block_number,
-        } = arguments;
-
-        let (state_root, block_number) = self
+        let parent_data = self.block.get_latest_block_hash_and_number();
+        let state_root = self
             .storage
-            .get_block_by_number(&block_number)
-            .map(|block| (block.header.state_root, block.header.number))
-            .unwrap_or_default();
-        debug!(
-            "Calling EVM at block number : {:#x}, state_root : {:#x}",
-            block_number, state_root
-        );
+            .get_latest_block()
+            .map_or(GENESIS_STATE_ROOT.parse().unwrap(), |block| {
+                block.header.state_root
+            });
 
-        let vicinity = Vicinity {
-            block_number,
-            origin: caller.unwrap_or_default(),
-            gas_limit: U256::from(gas_limit),
-            ..Default::default()
+        let (vicinity, parent_hash, current_block_number) = match parent_data {
+            None => (
+                Vicinity {
+                    beneficiary,
+                    timestamp: U256::from(timestamp),
+                    block_number: U256::zero(),
+                    ..Default::default()
+                },
+                H256::zero(),
+                U256::zero(),
+            ),
+            Some((hash, number)) => (
+                Vicinity {
+                    beneficiary,
+                    timestamp: U256::from(timestamp),
+                    block_number: number + 1,
+                    ..Default::default()
+                },
+                hash,
+                number + 1,
+            ),
         };
 
         let mut backend = EVMBackend::from_root(
             state_root,
-            Arc::clone(&self.trie_store),
+            Arc::clone(&self.core.trie_store),
             Arc::clone(&self.storage),
             vicinity,
-        )
-        .map_err(|e| anyhow!("------ Could not restore backend {}", e))?;
-        Ok(AinExecutor::new(&mut backend).call(ExecutorContext {
-            caller: caller.unwrap_or_default(),
-            to,
-            value,
-            data,
-            gas_limit,
-            access_list,
-        }))
-    }
-
-    pub fn validate_raw_tx(
-        &self,
-        tx: &str,
-        with_gas_usage: bool,
-    ) -> Result<(SignedTx, u64), Box<dyn Error>> {
-        debug!("[validate_raw_tx] raw transaction : {:#?}", tx);
-        let buffer = <Vec<u8>>::from_hex(tx)?;
-        let tx: TransactionV2 = ethereum::EnvelopedDecodable::decode(&buffer)
-            .map_err(|_| anyhow!("Error: decoding raw tx to TransactionV2"))?;
-        debug!("[validate_raw_tx] TransactionV2 : {:#?}", tx);
-
-        let block_number = self
-            .storage
-            .get_latest_block()
-            .map(|block| block.header.number)
-            .unwrap_or_default();
-
-        debug!("[validate_raw_tx] block_number : {:#?}", block_number);
-
-        let signed_tx: SignedTx = tx.try_into()?;
-        let nonce = self
-            .get_nonce(signed_tx.sender, block_number)
-            .map_err(|e| anyhow!("Error getting nonce {e}"))?;
-
-        debug!(
-            "[validate_raw_tx] signed_tx.sender : {:#?}",
-            signed_tx.sender
-        );
-        debug!(
-            "[validate_raw_tx] signed_tx nonce : {:#?}",
-            signed_tx.nonce()
-        );
-        debug!("[validate_raw_tx] nonce : {:#?}", nonce);
-
-        if nonce > signed_tx.nonce() {
-            return Err(anyhow!(
-                "Invalid nonce. Account nonce {}, signed_tx nonce {}",
-                nonce,
-                signed_tx.nonce()
-            )
-            .into());
-        }
-
-        const MIN_GAS_PER_TX: U256 = U256([21_000, 0, 0, 0]);
-        let balance = self
-            .get_balance(signed_tx.sender, block_number)
-            .map_err(|e| anyhow!("Error getting balance {e}"))?;
-
-        debug!("[validate_raw_tx] Accout balance : {:x?}", balance);
-
-        let prepay_gas = calculate_prepay_gas(&signed_tx);
-        debug!("[validate_raw_tx] prepay_gas : {:x?}", prepay_gas);
-
-        if balance < MIN_GAS_PER_TX || balance < prepay_gas {
-            debug!("[validate_raw_tx] insufficient balance to pay fees");
-            return Err(anyhow!("insufficient balance to pay fees").into());
-        }
-
-        let gas_limit = signed_tx.gas_limit();
-
-        debug!(
-            "[validate_raw_tx] MAX_GAS_PER_BLOCK: {:#x}",
-            MAX_GAS_PER_BLOCK
-        );
-        if gas_limit > MAX_GAS_PER_BLOCK {
-            return Err(anyhow!("Gas limit higher than MAX_GAS_PER_BLOCK").into());
-        }
-
-        let used_gas = if with_gas_usage {
-            let TxResponse { used_gas, .. } = self.call(EthCallArgs {
-                caller: Some(signed_tx.sender),
-                to: signed_tx.to(),
-                value: signed_tx.value(),
-                data: signed_tx.data(),
-                gas_limit: signed_tx.gas_limit().as_u64(),
-                access_list: signed_tx.access_list(),
-                block_number,
-            })?;
-            used_gas
-        } else {
-            u64::default()
-        };
-
-        Ok((signed_tx, used_gas))
-    }
-
-    pub fn logs_bloom(logs: Vec<Log>, bloom: &mut Bloom) {
-        for log in logs {
-            bloom.accrue(BloomInput::Raw(&log.address[..]));
-            for topic in log.topics {
-                bloom.accrue(BloomInput::Raw(&topic[..]));
-            }
-        }
-    }
-}
-
-// Transaction queue methods
-impl TxQueueService {
-    pub fn add_balance(
-        &self,
-        context: u64,
-        address: H160,
-        amount: U256,
-        hash: NativeTxHash,
-    ) -> Result<(), EVMError> {
-        let queue_tx = QueueTx::BridgeTx(BridgeTx::EvmIn(BalanceUpdate { address, amount }));
-        self.tx_queues.queue_tx(context, queue_tx, hash)?;
-        Ok(())
-    }
-
-    pub fn sub_balance(
-        &self,
-        context: u64,
-        address: H160,
-        amount: U256,
-        hash: NativeTxHash,
-    ) -> Result<(), EVMError> {
-        let block_number = self
-            .storage
-            .get_latest_block()
-            .map_or(U256::default(), |block| block.header.number);
-        let balance = self.get_balance(address, block_number)?;
-        if balance < amount {
-            Err(EVMBackendError::InsufficientBalance(InsufficientBalance {
-                address,
-                account_balance: balance,
-                amount,
-            })
-            .into())
-        } else {
-            let queue_tx = QueueTx::BridgeTx(BridgeTx::EvmOut(BalanceUpdate { address, amount }));
-            self.tx_queues.queue_tx(context, queue_tx, hash)?;
-            Ok(())
-        }
-    }
-
-    pub fn get_context(&self) -> u64 {
-        self.tx_queues.get_context()
-    }
-
-    pub fn clear(&self, context: u64) -> Result<(), EVMError> {
-        self.tx_queues.clear(context)?;
-        Ok(())
-    }
-
-    pub fn remove(&self, context: u64) {
-        self.tx_queues.remove(context);
-    }
-
-    pub fn remove_txs_by_sender(&self, context: u64, address: H160) -> Result<(), EVMError> {
-        self.tx_queues.remove_txs_by_sender(context, address)?;
-        Ok(())
-    }
-
-    /// Retrieves the next valid nonce for the specified account within a particular context.
-    ///
-    /// The method first attempts to retrieve the next valid nonce from the transaction queue associated with the
-    /// provided context. If no nonce is found in the transaction queue, that means that no transactions have been
-    /// queued for this account in this context. It falls back to retrieving the nonce from the storage at the latest
-    /// block. If no nonce is found in the storage (i.e., no transactions for this account have been committed yet),
-    /// the nonce is defaulted to zero.
-    ///
-    /// This method provides a unified view of the nonce for an account, taking into account both transactions that are
-    /// waiting to be processed in the queue and transactions that have already been processed and committed to the storage.
-    ///
-    /// # Arguments
-    ///
-    /// * `context` - The context queue number.
-    /// * `address` - The EVM address of the account whose nonce we want to retrieve.
-    ///
-    /// # Returns
-    ///
-    /// Returns the next valid nonce as a `U256`. Defaults to U256::zero()
-    pub fn get_next_valid_nonce_in_context(&self, context: u64, address: H160) -> U256 {
-        let nonce = self
-            .tx_queues
-            .get_next_valid_nonce(context, address)
-            .unwrap_or_else(|| {
-                let latest_block = self
-                    .storage
-                    .get_latest_block()
-                    .map(|b| b.header.number)
-                    .unwrap_or_else(U256::zero);
-
-                self.get_nonce(address, latest_block)
-                    .unwrap_or_else(|_| U256::zero())
-            });
-
-        debug!(
-            "Account {:x?} nonce {:x?} in context {context}",
-            address, nonce
-        );
-        nonce
-    }
-}
-
-// State methods
-impl TxQueueService {
-    pub fn get_account(
-        &self,
-        address: H160,
-        block_number: U256,
-    ) -> Result<Option<Account>, EVMError> {
-        let state_root = self
-            .storage
-            .get_block_by_number(&block_number)
-            .or_else(|| self.storage.get_latest_block())
-            .map(|block| block.header.state_root)
-            .unwrap_or_default();
-
-        let backend = EVMBackend::from_root(
-            state_root,
-            Arc::clone(&self.trie_store),
-            Arc::clone(&self.storage),
-            Vicinity::default(),
         )?;
-        Ok(backend.get_account(&address))
-    }
 
-    pub fn get_code(&self, address: H160, block_number: U256) -> Result<Option<Vec<u8>>, EVMError> {
-        self.get_account(address, block_number).map(|opt_account| {
-            opt_account.map_or_else(
-                || None,
-                |account| self.storage.get_code_by_hash(account.code_hash),
-            )
-        })
-    }
+        let mut executor = AinExecutor::new(&mut backend);
 
-    pub fn get_storage_at(
-        &self,
-        address: H160,
-        position: U256,
-        block_number: U256,
-    ) -> Result<Option<Vec<u8>>, EVMError> {
-        self.get_account(address, block_number)?
-            .map_or(Ok(None), |account| {
-                let storage_trie = self
-                    .trie_store
-                    .trie_db
-                    .trie_restore(address.as_bytes(), None, account.storage_root.into())
-                    .unwrap();
+        for (queue_tx, hash) in self.core.tx_queues.get_cloned_vec(context) {
+            match queue_tx {
+                QueueTx::SignedTx(signed_tx) => {
+                    if ain_cpp_imports::past_changi_intermediate_height_4_height() {
+                        let nonce = executor.get_nonce(&signed_tx.sender);
+                        if signed_tx.nonce() != nonce {
+                            return Err(anyhow!("EVM block rejected for invalid nonce. Address {} nonce {}, signed_tx nonce: {}", signed_tx.sender, nonce, signed_tx.nonce()).into());
+                        }
+                    }
 
-                let tmp: &mut [u8; 32] = &mut [0; 32];
-                position.to_big_endian(tmp);
-                storage_trie
-                    .get(tmp.as_slice())
-                    .map_err(|e| EVMError::TrieError(format!("{e}")))
-            })
-    }
+                    let (
+                        TxResponse {
+                            exit_reason,
+                            logs,
+                            used_gas,
+                            ..
+                        },
+                        receipt,
+                    ) = executor.exec(&signed_tx);
+                    debug!(
+                        "receipt : {:#?} for signed_tx : {:#x}",
+                        receipt,
+                        signed_tx.transaction.hash()
+                    );
 
-    pub fn get_balance(&self, address: H160, block_number: U256) -> Result<U256, EVMError> {
-        let balance = self
-            .get_account(address, block_number)?
-            .map_or(U256::zero(), |account| account.balance);
+                    if !exit_reason.is_succeed() {
+                        failed_transactions.push(hex::encode(hash));
+                    }
 
-        debug!("Account {:x?} balance {:x?}", address, balance);
-        Ok(balance)
-    }
+                    all_transactions.push(signed_tx.clone());
+                    gas_used += used_gas;
+                    EVMCoreService::logs_bloom(logs, &mut logs_bloom);
+                    receipts_v3.push(receipt);
+                }
+                QueueTx::BridgeTx(BridgeTx::EvmIn(BalanceUpdate { address, amount })) => {
+                    debug!(
+                        "[finalize_block] EvmIn for address {:x?}, amount: {}, context {}",
+                        address, amount, context
+                    );
+                    if let Err(e) = executor.add_balance(address, amount) {
+                        debug!("[finalize_block] EvmIn failed with {e}");
+                        failed_transactions.push(hex::encode(hash));
+                    }
+                }
+                QueueTx::BridgeTx(BridgeTx::EvmOut(BalanceUpdate { address, amount })) => {
+                    debug!(
+                        "[finalize_block] EvmOut for address {}, amount: {}",
+                        address, amount
+                    );
 
-    pub fn get_nonce(&self, address: H160, block_number: U256) -> Result<U256, EVMError> {
-        let nonce = self
-            .get_account(address, block_number)?
-            .map_or(U256::zero(), |account| account.nonce);
-
-        debug!("Account {:x?} nonce {:x?}", address, nonce);
-        Ok(nonce)
-    }
-
-    pub fn get_latest_block_backend(&self) -> Result<EVMBackend, EVMBackendError> {
-        let (state_root, block_number) = self
-            .storage
-            .get_latest_block()
-            .map(|block| (block.header.state_root, block.header.number))
-            .unwrap_or_default();
-
-        debug!(
-            "[get_latest_block_backend] At block number : {:#x}, state_root : {:#x}",
-            block_number, state_root
-        );
-        EVMBackend::from_root(
-            state_root,
-            Arc::clone(&self.trie_store),
-            Arc::clone(&self.storage),
-            Default::default(),
-        )
-    }
-}
-
-use std::fmt;
-
-#[derive(Debug)]
-pub enum EVMError {
-    BackendError(EVMBackendError),
-    QueueError(QueueError),
-    NoSuchAccount(H160),
-    TrieError(String),
-}
-
-impl fmt::Display for EVMError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            EVMError::BackendError(e) => write!(f, "EVMError: Backend error: {e}"),
-            EVMError::QueueError(e) => write!(f, "EVMError: Queue error: {e}"),
-            EVMError::NoSuchAccount(address) => {
-                write!(f, "EVMError: No such acccount for address {address:#x}")
+                    if let Err(e) = executor.sub_balance(address, amount) {
+                        debug!("[finalize_block] EvmOut failed with {e}");
+                        failed_transactions.push(hex::encode(hash));
+                    }
+                }
             }
-            EVMError::TrieError(e) => {
-                write!(f, "EVMError: Trie error {e}")
-            }
+
+            executor.commit();
         }
+
+        if update_state {
+            self.core.tx_queues.remove(context);
+        }
+
+        let block = Block::new(
+            PartialHeader {
+                parent_hash,
+                beneficiary,
+                state_root: if update_state {
+                    backend.commit()
+                } else {
+                    backend.root()
+                },
+                receipts_root: ReceiptService::get_receipts_root(&receipts_v3),
+                logs_bloom,
+                difficulty: U256::from(difficulty),
+                number: current_block_number,
+                gas_limit: MAX_GAS_PER_BLOCK,
+                gas_used: U256::from(gas_used),
+                timestamp,
+                extra_data: Vec::default(),
+                mix_hash: H256::default(),
+                nonce: H64::default(),
+            },
+            all_transactions
+                .iter()
+                .map(|signed_tx| signed_tx.transaction.clone())
+                .collect(),
+            Vec::new(),
+        );
+
+        let receipts = self.receipt.generate_receipts(
+            &all_transactions,
+            receipts_v3,
+            block.header.hash(),
+            block.header.number,
+        );
+
+        if update_state {
+            debug!(
+                "[finalize_block] Finalizing block number {:#x}, state_root {:#x}",
+                block.header.number, block.header.state_root
+            );
+            // calculate base fee
+            let base_fee = self.block.calculate_base_fee(parent_hash);
+
+            self.block.connect_block(block.clone(), base_fee);
+            self.logs
+                .generate_logs_from_receipts(&receipts, block.header.number);
+            self.receipt.put_receipts(receipts);
+            self.filters.add_block_to_filters(block.header.hash());
+        }
+
+        Ok((
+            *block.header.hash().as_fixed_bytes(),
+            failed_transactions,
+            gas_used,
+        ))
+    }
+
+    pub fn queue_tx(&self, context: u64, tx: QueueTx, hash: NativeTxHash) -> Result<(), EVMError> {
+        self.core.tx_queues.queue_tx(context, tx.clone(), hash)?;
+
+        if let QueueTx::SignedTx(signed_tx) = tx {
+            self.filters.add_tx_to_filters(signed_tx.transaction.hash())
+        }
+
+        Ok(())
     }
 }
-
-impl From<EVMBackendError> for EVMError {
-    fn from(e: EVMBackendError) -> Self {
-        EVMError::BackendError(e)
-    }
-}
-
-impl From<QueueError> for EVMError {
-    fn from(e: QueueError) -> Self {
-        EVMError::QueueError(e)
-    }
-}
-
-impl std::error::Error for EVMError {}
