+++ lib/ain-evm/src/lib.rs
@@ -2,18 +2,18 @@ mod backend;
 pub mod block;
 pub mod bytes;
 mod ecrecover;
-pub mod evm;
+pub mod core;
 pub mod executor;
 mod fee;
 pub mod filters;
 mod genesis;
-pub mod handler;
+pub mod evm;
 pub mod log;
 mod precompiles;
 pub mod receipt;
-pub mod runtime;
+pub mod services;
 pub mod storage;
 pub mod traits;
 pub mod transaction;
 mod trie;
-mod tx_queue;
+mod txqueue;
