+++ lib/ain-evm/src/services.rs
@@ -1,4 +1,4 @@
-use crate::handler::EVMServices;
+use crate::evm::EVMServices;
 use crate::storage::traits::FlushableStorage;
 
 use jsonrpsee_http_server::HttpServerHandle;
@@ -47,13 +47,6 @@ impl Services {
 
     pub fn stop(&self) {
         let _ = self.tokio_runtime_channel_tx.blocking_send(());
-        self.json_rpc
-            .lock()
-            .unwrap()
-            .take()
-            .unwrap()
-            .stop()
-            .unwrap();
 
         self.tokio_worker
             .lock()
@@ -64,7 +57,7 @@ impl Services {
             .unwrap();
 
         // Persist EVM State to disk
-        self.evm.queue.flush().expect("Could not flush evm state");
+        self.evm.core.flush().expect("Could not flush evm state");
         self.evm.storage.flush().expect("Could not flush storage");
     }
 }
