+++ lib/ain-grpc/src/lib.rs
@@ -36,7 +36,7 @@ use std::error::Error;
 use std::net::SocketAddr;
 use std::sync::Arc;
 
-use ain_evm::runtime::{Services, SERVICES};
+use ain_evm::services::{Services, SERVICES};
 
 pub fn preinit() {}
